refactor: change file structure
tmp3
